--- src/custom/utils/analytics/actions/index.ts
@@ -1,5 +1,0 @@
-export * from './lists'
-export * from './settings'
-export * from './theme'
-export * from './transactions'
-export * from './wallet'
