+++ src/custom/utils/analytics/settingsEvents.ts
@@ -1,5 +1,4 @@
-import { Category } from '../types'
-import { _reportEvent } from '../utils'
+import { Category, _reportEvent } from './index'
 import { debounce } from 'utils/misc'
 
 const types = {
