--- src/custom/utils/analytics/types.ts
@@ -1,20 +1,0 @@
-export interface EventParams {
-  category: Category
-  action: string
-  label?: string
-  value?: number
-}
-
-export enum Category {
-  SWAP = 'Swap',
-  LIST = 'Lists',
-  CURRENCY_SELECT = 'Currency Select',
-  EXPERT_MODE = 'Expert mode',
-  RECIPIENT_ADDRESS = 'Recipient address',
-  ORDER_SLIPAGE_TOLERANCE = 'Order Slippage Tolerance',
-  ORDER_EXPIRATION_TIME = 'Order Expiration Time',
-  WALLET = 'Wallet',
-  WRAP_NATIVE_TOKEN = 'Wrapped Native Token',
-  CLAIM_COW_FOR_LOCKED_GNO = 'Claim COW for Locked GNO', // TODO: Maybe Claim COW was enough?
-  THEME = 'Theme',
-}
